+++ konduit-serving-data/konduit-serving-image/src/main/java/ai/konduit/serving/data/image/convert/config/AspectRatioHandling.java
@@ -1,0 +1,5 @@
+package ai.konduit.serving.data.image.convert.config;
+
+public enum AspectRatioHandling {
+    CENTER_CROP, PAD, STRETCH
+}
