+++ konduit-serving-data/konduit-serving-image/src/main/java/ai/konduit/serving/data/image/convert/config/NDDepth.java
@@ -1,0 +1,5 @@
+package ai.konduit.serving.data.image.convert.config;
+
+public enum NDDepth {
+    RGB, RGBA, BGR, BGRA, GRAYSCALE
+}
