+++ konduit-serving-data/konduit-serving-image/src/main/java/ai/konduit/serving/data/image/convert/config/NDFormat.java
@@ -1,0 +1,6 @@
+package ai.konduit.serving.data.image.convert.config;
+
+public enum NDFormat {
+    CHANNELS_FIRST,
+    CHANNELS_LAST
+}
